+++ pnet-data-api-java/src/main/java/pnet/data/api/person/PersonType.java
@@ -8,6 +8,11 @@ public enum PersonType
     PERSON,
 
     /**
+     * A privileged (admin) user directly related to a regular person.
+     */
+    PRIVILEGED,
+
+    /**
      * A user for automated tasks, e.g. for automatically executed tests.
      */
     BOT,
@@ -15,5 +20,10 @@ public enum PersonType
     /**
      * A user for manual tests.
      */
-    TEST_USER
+    TEST_USER,
+
+    /**
+     * A user for manual tests with privileged rights.
+     */
+    PRIVILEGED_TEST_USER
 }